fix(job): successful callback was reported as error (#13)
Before this fix, the job with the callback, would make successful request,
receive the result and report this result as error. That is because in the
promise .then chain, done callback was passed as is. Done callback accepted
error as a first parameter, but was receiving request result in place of error.